Fixes fatal error if $post is empty
In some cases, the $post variable is not existent when this code is run, causing the whole metabox to have a fatal error and the order can not be edited at all.

This fix wraps everything that happens after `global $post;` in an (!empty()) function, so if $post is empty, nothing is run.

This is a temporary patch, we really should give a graceful error status, but I'm not 100% sure on the full functionality of the box, hence this fix.