Fixing CurrencyCode bug (#211)
For upstream price sync, we can't specify the CurrencyCode attribute, becaue apparantly, it stands for the procurement currency paid by the vendor, but not the currency the product is sold in.

This means that if the purchasing currency for a product is 500 EUR originally and the upstream price sync changes the code to ISK, the value changes to 500 ISK. That in turn affects margin calculations for any sale of the product substantially.

Removing this single line of code may solve the issue but this may also require some further research.